Rename `DapClient.ts` to `DapComm.ts`
